feat: cleanup background utils
