+++ src/scripts/background/icon.ts
